Install ost deps as well
virtualenv is missing when installing lago using this script.
This means that the invocation using the suite name should fail as well
on scratch centos machine.

Added a function to resolve all those deps.

TODO a dependencies.yaml in ost would be nice to resolve or just a list
in file

Signed-off-by: Roy Golan <rgolan@redhat.com>